Adding book trail and making worker details modular with separate class component for complete and trial booking
